pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Python scripts"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "mdtoc>=1.3",
    "pyyaml>=6.0.2",
]
